Revert "Move transformers out of class exclusion"
This reverts commit b87317e730bf1d4c00cda30ef55cd12400eb731e.
