+++ src/main/java/org/bookmc/loader/api/launch/transform/QuiltRemapper.java
@@ -1,4 +1,4 @@
-package org.bookmc.external.transformer;
+package org.bookmc.loader.api.launch.transform;
 
 /**
  * A specialised transformer for remapping the game at runtime
