+++ src/main/java/org/bookmc/loader/api/launch/transform/QuiltTransformer.java
@@ -1,4 +1,4 @@
-package org.bookmc.external.transformer;
+package org.bookmc.loader.api.launch.transform;
 
 public interface QuiltTransformer {
     byte[] transform(String name, byte[] clazz);
