+++ src/main/java/org/bookmc/loader/impl/Loader.java
@@ -9,8 +9,8 @@ import org.bookmc.loader.api.candidate.ModCandidate;
 import org.bookmc.loader.api.classloader.IQuiltClassLoader;
 import org.bookmc.loader.api.classloader.ModClassLoader;
 import org.bookmc.loader.api.exception.IllegalDependencyException;
-import org.bookmc.external.transformer.QuiltRemapper;
-import org.bookmc.external.transformer.QuiltTransformer;
+import org.bookmc.loader.api.launch.transform.QuiltRemapper;
+import org.bookmc.loader.api.launch.transform.QuiltTransformer;
 import org.bookmc.loader.api.vessel.ModVessel;
 import org.bookmc.loader.api.vessel.dependency.ModDependency;
 import org.bookmc.loader.api.vessel.entrypoint.Entrypoint;
